Add solution for the course scheduling problem with better complexity
Implemented a new function 'canFinish5' to solve the LeetCode problem - 207 Course Schedule. The new function uses the graph theory and a topological sorting algorithm to find whether it is possible to finish all courses. The solution solves the problem by checking if the graph of prerequisite pairs has a cycle or not with a better complexity.