[4] 双指针, 为了节省内存..? Add algorithm to count pairs in range
This commit adds a new file (lc1782_countPairs.go) which implements an algorithm to count pairs in a specified range ('queries'). The algorithm calculates degrees of nodes in a given edge list and then uses these degrees to determine and sum up all pairs whose total degree is greater than each query. It uses sorting and a two-pointer technique to consume less time, and map and slice data structures to save memory.